refactor(loops): move tree projection to loop_tree.go
Keeps loop_view.go focused on the LoopView row projection and under the
500-line large-file threshold; the tree projection is a distinct concern.
No behavior change.

Refs #1102
